+++ server/data-service/src/main/java/edu/illinois/ncsa/incore/service/data/controllers/DatasetController.java
@@ -488,15 +488,7 @@ public class DatasetController {
                 }
                 // remove geoserver layer
                 if (geoserverUsed) {
-                    if (format.equalsIgnoreCase(FileUtils.FORMAT_NETWORK)) {
-                        // remove network dataset
-                        boolean linkRemoved = GeoserverUtils.removeLayerFromGeoserver(datasetId, "_link");
-                        boolean nodeRemoved = GeoserverUtils.removeLayerFromGeoserver(datasetId, "_node");
-                        boolean storeRemoved = GeoserverUtils.removeStoreFromGeoserver(datasetId);
-                    } else {
-                        boolean layerRemoved = GeoserverUtils.removeLayerFromGeoserver(datasetId);
-                        boolean storeRemoved = GeoserverUtils.removeStoreFromGeoserver(datasetId);
-                    }
+                    boolean isRemoved = GeoserverUtils.removeStoreFromGeoserver(datasetId);
                 }
             }
         } else {
@@ -623,6 +615,7 @@ public class DatasetController {
         boolean isTif = false;
         boolean isZip = false;
         boolean isJoin = false;
+        boolean isPrj = false;
 
         int fileCounter = 0;
         int linkCounter = 0;
@@ -649,6 +642,8 @@ public class DatasetController {
                     isShp = true;
                 } else if (fileExt.equalsIgnoreCase("zip")) {
                     isZip = true;
+                } else if (fileExt.equalsIgnoreCase("prj")) {
+                    isPrj = true;
                 }
 
                 // process zip file
@@ -762,7 +757,6 @@ public class DatasetController {
         List<File> files = new ArrayList<File>();
 
         File geoPkgFile = null;
-        boolean isShpfile = false;
 
         if (format.equalsIgnoreCase(FileUtils.FORMAT_SHAPEFILE) || format.equalsIgnoreCase(FileUtils.FORMAT_NETWORK)) {
             for (int i = 0; i < dataFDs.size(); i++) {
@@ -773,7 +767,9 @@ public class DatasetController {
                 //get file, if the file is in remote, use http downloader
                 String fileExt = FilenameUtils.getExtension(shpLoc);
                 if (fileExt.equalsIgnoreCase(FileUtils.EXTENSION_SHP)) {
-                    isShpfile = true;
+                    isShp = true;
+                } else if (fileExt.equalsIgnoreCase(FileUtils.EXTENSION_PRJ)) {
+                    isPrj = true;
                 }
             }
             try {
@@ -804,6 +800,12 @@ public class DatasetController {
                             try {
                                 InputStream is = new FileInputStream(shpFile);
                                 String fileName = shpFile.getName();
+                                String fileExt = FilenameUtils.getExtension(shpFile.getName());
+                                if (fileExt.equalsIgnoreCase(FileUtils.EXTENSION_PRJ)) {
+                                    isPrj = true;
+                                } else if (fileExt.equalsIgnoreCase(FileUtils.EXTENSION_SHP)) {
+                                    isShp = true;
+                                }
                                 fd = fsDisk.storeFile(fileName, is);
                                 fd.setFilename(fileName);
                             } catch (IOException e) {
@@ -819,9 +821,6 @@ public class DatasetController {
                         logger.debug("Unzipping zip file failed");
                         throw new IncoreHTTPException(Response.Status.INTERNAL_SERVER_ERROR, "Unzipping zip file failed.");
                     }
-
-                    // if the unzip was successful, that means that it is a complete shapefile
-                    isShp = true;
 
                     // create and add FileDescriptor for unzipped shapefiles
                     List<FileDescriptor> unzippedFDs = dataset.getFileDescriptors();
@@ -896,9 +895,20 @@ public class DatasetController {
             } else {
                 try {
                     if (format.equalsIgnoreCase(FileUtils.FORMAT_NETWORK)) {
-                        GeoserverUtils.networkDatasetUploadToGeoserver(dataset, repository);
-                    } else {
+                        if (isShp && isPrj) {
+                            GeoserverUtils.networkDatasetUploadToGeoserver(dataset, repository);
+                        } else {
+                            logger.error("There is no prj file. Uploading to geoserver has been canceled");
+                        }
+                    } else if (format.equalsIgnoreCase("raster") || format.equalsIgnoreCase("geotiff") ||
+                        format.equalsIgnoreCase("tif") || format.equalsIgnoreCase("tiff")) {
                         GeoserverUtils.datasetUploadToGeoserver(dataset, repository, isShp, isTif, isAsc);
+                    } else {
+                        if (isShp && isPrj) {
+                            GeoserverUtils.datasetUploadToGeoserver(dataset, repository, isShp, isTif, isAsc);
+                        } else {
+                            logger.error("There is no prj file. Uploading to geoserver has been canceled");
+                        }
                     }
                 } catch (IOException e) {
                     logger.error("Error uploading dataset to geoserver ", e);